Adding transposition file to xcode project and fixing precision warning

+++ src/transposition.cpp
@@ -664,7 +664,7 @@ TransPitch Transposer::IntegerToPitch(int ipitch)
         // search from B downwards
         mindiff = chroma - m_diatonicMapping.back();
         mini = (int)m_diatonicMapping.size() - 1;
-        for (int i = m_diatonicMapping.size() - 2; i >= 0; i--) {
+        for (int i = (int)m_diatonicMapping.size() - 2; i >= 0; i--) {
             int diff = chroma - m_diatonicMapping[i];
             if (abs(diff) < abs(mindiff)) {
                 mindiff = diff;
